requirements.txt
vanna[sqlite]>=2.0.0
fastapi
uvicorn[standard]
plotly
pandas
python-dotenv
openai
groq
slowapi
httpx
